Exclude missing questions only for certain indicators
Following the logic for other indicators (quant) it seems that the aggregate score excludes countries where no response has been provided.  But this provides a distortionary picture too!  Eg GAVI for 7DP gets a score of 100% which looks like they participate in mutual assessments wherever they exist.  But that's not true.  So we need to derive the denominator from a combination of a) countries in which the agency is reporting; b) countries that have resported to Yes to 1G, 6G or 7G.